Delete removed annotations from DateTime constants
PhpStorm marks code using these constants as broken as they have been removed. But they can still be used as they are inherited from the DateTimeInterface.
In theory we could remove the constants from DateTime completely. But the interface constants have `since 7.2` which would also be wrong for DateTime constants.
So I think this is the best solution.

Ref. #951