Flag functions that are marked pure that depend on external state
Adds the `$mayDependOnGlobalScope` to functions which depend on
external state. There are a multitude of reasons for this:

* Uses the system clock
* Reads system configuration
* Influenced by PHP ini settings
* Locale dependant
* Reads from the filesystem
* Based on PHP symbols/variable scope

Initially this started out as just adding the flag to `hrtime()` but
I noticed so many other cass in the same file I went through
everything marked pure in the `standard/` and `Core/` directories.